services/gateway/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
email-validator==2.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database & Cache
redis==5.0.1
redis[hiredis]==5.0.1

# HTTP client
httpx[http2]==0.25.2

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Utilities
python-dotenv==1.0.0

# Performance optimizations
uvloop==0.19.0
httptools==0.6.1
